BUGFIX: .selfcontrol config creation.
Grrr. way back in time during cleanup, removed the bit that returns
a default config if no ~/.selfconfig file was found.  Removed it from
bin/SelfControl and forgot to move it to lib/SelfControl/Config.pm

Added to Config.pm with defaults suitable for version 2 of the config
file format.